# docs/contracts/openapi-v1.yaml
...
openapi: 3.0.3
info:
  title: CTF Platform API (v1)
  version: 1.0.0
  description: '机器可读接口契约（OpenAPI 3.0）。


    - 业务字段/返回结构以 `ctf/docs/contracts/api-contract-v1.md` 为准

    - 错误码以 `ctf/docs/architecture/backend/04-api-design.md` §3 为准

    - WebSocket 端点与消息格式见 `ctf/docs/contracts/api-contract-v1.md` §9（不在 OpenAPI 范围）

    '
  x-last-updated: '2026-05-14'
servers:
- {url: /api/v1}
...
components:
  securitySchemes:
    sessionCookieAuth: {type: apiKey, in: cookie, name: ctf_session}
  schemas:
...
    ISODateTime: {type: string, format: date-time, description: 'RFC3339/ISO8601，服务端统一以 UTC 输出，如 "2026-03-01T03:15:22Z"'}
    FieldError:
      type: object
      required: [field, message]
      properties: {field: {type: string}, message: {type: string}}
    ApiEnvelopeBase:
      type: object
      required: [code, message, request_id]
      properties:
        code: {type: integer, description: 0=成功，非0=业务错误码（见 backend/04-api-design.md §3）}
        message: {type: string, description: 成功固定 "success"，失败为摘要信息（前端提示优先走 errorMap）}
        request_id: {type: string, description: 必填：用于日志追踪}
        errors: {type: array, items: {$ref: '#/components/schemas/FieldError'}}
    ApiEnvelopeNull:
      allOf:
      - {$ref: '#/components/schemas/ApiEnvelopeBase'}
      - type: object
        required: [data]
        properties: {data: {nullable: true, description: 无业务数据时为 null}}
...
    MCPTokenData:
      type: object
      required: [token, expires_at]
      properties: {token: {type: string}, expires_at: {$ref: '#/components/schemas/ISODateTime'}}
    ApiEnvelopeMCPTokenData:
      allOf:
      - {$ref: '#/components/schemas/ApiEnvelopeBase'}
      - type: object
        required: [data]
        properties: {data: {$ref: '#/components/schemas/MCPTokenData'}}
...
paths:
...
  /auth/mcp-token:
    post:
      tags: [Auth]
      summary: 获取外部 agent 调用 MCP 的 Bearer token
      security:
      - {sessionCookieAuth: []}
      responses:
        '200':
          description: OK
          content:
            application/json: {schema: {$ref: '#/components/schemas/ApiEnvelopeMCPTokenData'}}
        '401':
          description: Unauthorized
          content:
            application/json: {schema: {$ref: '#/components/schemas/ApiEnvelopeNull'}}
...